refactor : [코드리뷰 반영] 테스트코드 패키지
